+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
   and budget pages
 - Added Meredith Fuchs to Leadership calendar filter.
 - Added unit test for `assign` utility.
+- Added `get-breakpoint-state.js` to add support for responsive JS.
 
 ### Changed
 - Moved `.meta-header`, `.jump-link`,
@@ -64,6 +65,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Normalized director and deputy director photos to be format `NAME-WxH.jpg`.
 - Changed name of `shallow-extend` utility to 'assign'.
 - Superscripts `st` in `21st` on About Us page.
+- Updated `BreakpointHandler.js` to support usage of `breakpoints-config.js`.
 
 ### Removed
 - Removed styles from codebase that have already been migrated